+++ src/index.ts
@@ -1,4 +1,4 @@
-export const version = "0.17.0";
+export const version = "0.18.0";
 
 export { cn } from "./lib/cn";
 export { useMediaQuery, useMobile } from "./lib/screen";